gke-automation/requirements.txt
aiohttp
pyyaml
google-auth
oauthlib
requests
flask
gunicorn